Feat : 게시물 및 댓글 CRUD 추가

+++ src/main/java/com/stamfee/stamfee/controller/CommunityRestController.java
@@ -14,7 +14,9 @@ import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.BindParam;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -40,51 +42,51 @@ public class CommunityRestController {
     int pageSize;
 
     @PostMapping("/addPost")
-    public ResponseEntity<String> addPost(@RequestPart PostDTO postDTO, @RequestParam(value = "images", required = false) List<MultipartFile> images) throws Exception {
-        PostDTO post = postService.addPost(postDTO);
+    public ResponseEntity<Long> addPost(@ModelAttribute PostDTO postDTO, @RequestParam(value = "image", required = false) List<MultipartFile> image) throws Exception {
+        log.info("받은 post는 : {}",postDTO);
+        Long postId = postService.addPost(postDTO);
 
         // 이미지가 제공되었는지 확인
-        if (images != null && !images.isEmpty()) {
+        if (image != null && !image.isEmpty()) {
             // 이미지 업로드 및 저장을 위한 ImageDTO 생성
             ImageDTO imageDTO = ImageDTO.builder()
-                    .post(PostDTO.builder().postId(post.getPostId()).build())
+                    .post(PostDTO.builder().postId(postId).build())
                     .build();
 
             // 이미지 업로드 및 저장
-            imageService.addImage(images, imageDTO);
+            imageService.addImage(image, imageDTO);
         }
-        return ResponseEntity.status(HttpStatus.OK).body("{\"success\": true, \"postId\": %d}\", postId}");
+        return postId != null ? ResponseEntity.ok(postId) : ResponseEntity.badRequest().build();
     }
 
     @GetMapping("/getPost/{postId}")
-    public PostDTO getPost(@PathVariable long postId) throws Exception {
+    public ResponseEntity<PostDTO> getPost(@PathVariable long postId) throws Exception {
+        log.info("받은 postId는 : {}",postId);
         List<ImageDTO> imageDTOList =imageService.getImageList(postId, SearchDTO.builder()
-                .isPost(true)
                 .build());
         PostDTO postDTO = postService.getPost(postId);
         List<String> imageNames = imageDTOList.stream()
                 .map(ImageDTO::getImageName)
                 .collect(Collectors.toList());
         postDTO.setImages(imageNames);
-        return postDTO;
+        return postDTO != null ? ResponseEntity.ok(postDTO) : ResponseEntity.badRequest().build();
     }
 
 
-    @GetMapping("/dongNeHome/{currentPage}")
-    public List<PostDTO> getPostList(@PathVariable int currentPage, @RequestParam String gu, @RequestParam(required = false) String searchKeyword) throws Exception {
-        System.out.println(gu);
+    @GetMapping("/getPostList/{currentPage}")
+    public ResponseEntity<List<PostDTO>> getPostList(@PathVariable int currentPage, @RequestParam(value = "searchKeyword",required = false) String searchKeyword) throws Exception {
         System.out.println(searchKeyword);
         SearchDTO searchDTO = SearchDTO.builder()
                 .currentPage(currentPage)
                 .pageSize(pageSize)
                 .searchKeyword(searchKeyword)
                 .build();
-
-        return postService.getPostList(searchDTO);
+        List<PostDTO> postDTOList = postService.getPostList(searchDTO);
+        return postDTOList != null ? ResponseEntity.ok(postDTOList) : ResponseEntity.badRequest().build();
     }
 
     @PostMapping("/updatePost")
-    public ResponseEntity<String> updatePost(@RequestBody PostDTO postDTO,@RequestParam("images") List<MultipartFile> images) throws Exception {
+    public ResponseEntity<Long> updatePost(@ModelAttribute PostDTO postDTO,@RequestParam("image") List<MultipartFile> image) throws Exception {
         Long postId = postService.updatePost(postDTO);
         imageService.deletePostImage(postId);
 
@@ -94,43 +96,49 @@ public class CommunityRestController {
             .build();
 
         // 이미지 업로드 및 저장
-        imageService.addImage(images, imageDTO);
-        return ResponseEntity.status(HttpStatus.OK).body("{\"success\": true}");
+        imageService.addImage(image, imageDTO);
+        return postId != null ? ResponseEntity.ok(postId) : ResponseEntity.badRequest().build();
     }
 
     @GetMapping("/deletePost/{postId}")
-    public ResponseEntity<String> deletePost(@PathVariable long postId) throws Exception {
-        postService.deletePost(postId);
-        imageService.deletePostImage(postId);
-        return ResponseEntity.status(HttpStatus.OK).body("{\"success\": true}");
+    public ResponseEntity<Boolean> deletePost(@PathVariable long postId) throws Exception {
+        if(postService.deletePost(postId)){
+            imageService.deletePostImage(postId);
+            return ResponseEntity.ok(true);
+        }else
+            return ResponseEntity.badRequest().body(false);
     }
 
     @PostMapping("/addComment")
-    public ResponseEntity<String> addComment(@RequestBody CommentDTO commentDTO) throws Exception {
-        commentService.addComment(commentDTO);
-        return ResponseEntity.status(HttpStatus.OK).body("{\"success\": true}");
+    public ResponseEntity<Boolean> addComment(@RequestBody CommentDTO commentDTO) throws Exception {
+        boolean result = commentService.addComment(commentDTO);
+        return result ? ResponseEntity.ok(true): ResponseEntity.badRequest().body(false);
     }
 
     @GetMapping("/getComment/{commentId}")
-    public CommentDTO getComment(@PathVariable long commentId) throws Exception {
-        return commentService.getComment(commentId);
+    public ResponseEntity<CommentDTO> getComment(@PathVariable long commentId) throws Exception {
+
+        CommentDTO commentDTO = commentService.getComment(commentId);
+
+        return commentDTO != null ? ResponseEntity.ok(commentDTO) : ResponseEntity.badRequest().build();
     }
 
     @GetMapping("/getCommentList/{postId}")
-    public List<CommentDTO> getCommentList(@PathVariable  long postId) throws Exception {
-        return commentService.getCommentList(postId);
+    public ResponseEntity<List<CommentDTO>> getCommentList(@PathVariable  long postId) throws Exception {
+        List<CommentDTO> commentDTOList = commentService.getCommentList(postId);
+        return commentDTOList != null ? ResponseEntity.ok(commentDTOList) : ResponseEntity.badRequest().build();
     }
 
     @PostMapping("/updateComment")
-    public ResponseEntity<String> updateComment(@RequestBody CommentDTO commentDTO) throws Exception {
-        commentService.updateComment(commentDTO);
-        return ResponseEntity.status(HttpStatus.OK).body("{\"success\": true}");
+    public ResponseEntity<Boolean> updateComment(@RequestBody CommentDTO commentDTO) throws Exception {
+        boolean result = commentService.updateComment(commentDTO);
+        return result ? ResponseEntity.ok(true) : ResponseEntity.badRequest().body(false);
     }
 
     @GetMapping("/deleteComment/{commentId}")
-    public ResponseEntity<String> deleteComment(@PathVariable long commentId) throws Exception {
-        commentService.deleteComment(commentId);
-        return ResponseEntity.status(HttpStatus.OK).body("{\"success\": true}");
+    public ResponseEntity<Boolean> deleteComment(@PathVariable long commentId) throws Exception {
+        boolean result = commentService.deleteComment(commentId);
+        return result ? ResponseEntity.ok(true) : ResponseEntity.badRequest().body(false);
     }
 
 }
